+++ ApexDAG/experiments/configs/finetune/default.yaml
@@ -8,7 +8,9 @@ edge_classes: 8
 batch_size: 32
 learning_rate: 0.001
 num_epochs: 100
-train_split: 0.8
+train_split: 0.6
+test_split: 0.2
+val_split: 0.2
 patience: 10
 
 # hyperparameters for encoder
