+++ ApexDAG/nn/training/base_trainer.py
@@ -1,0 +1,150 @@
+import os
+import tqdm
+import wandb
+import torch
+import torch.nn as nn
+import torch.optim as optim
+import matplotlib.pyplot as plt
+import seaborn as sns
+
+from datetime import datetime
+from torch_geometric.loader import DataLoader
+from torch.utils.tensorboard import SummaryWriter
+from sklearn.metrics import confusion_matrix
+
+global INTERRUPTED
+INTERRUPTED = False
+
+class BaseTrainer:
+    def __init__(self, model, train_dataset, val_dataset, device="cpu", log_dir="runs/", checkpoint_dir="checkpoints/", patience=10):
+        self.model = model.to(device)
+        self.device = device
+        self.train_loader = DataLoader(train_dataset, batch_size=32, shuffle=True)
+        self.val_loader = DataLoader(val_dataset, batch_size=32, shuffle=False)
+        self.optimizer = optim.Adam(self.model.parameters(), lr=0.001)
+        
+        self.criterion_node = nn.CrossEntropyLoss()
+        self.criterion_edge_type = nn.CrossEntropyLoss()
+        self.criterion_edge_existence = nn.BCELoss()
+        
+        self.writer = SummaryWriter(log_dir=log_dir)
+        self.checkpoint_dir = checkpoint_dir
+        os.makedirs(checkpoint_dir, exist_ok=True)
+
+        self.patience = patience
+        self.best_val_loss = float("inf")
+        self.early_stopping_counter = 0
+
+    def save_checkpoint(self, epoch, val_loss, filename=None):
+        if filename is None:
+            filename = f"model_epoch_{epoch}.pt"
+        checkpoint_path = os.path.join(self.checkpoint_dir, filename)
+        torch.save({
+            'epoch': epoch,
+            'model_state_dict': self.model.state_dict(),
+            'optimizer_state_dict': self.optimizer.state_dict(),
+            'val_loss': val_loss
+        }, checkpoint_path)
+
+    def log_confusion_matrix(self, loader, phase):
+        self.model.eval()
+        all_preds = []
+        all_labels = []
+        
+        with torch.no_grad():
+            for data in loader:
+                data = data.to(self.device)
+                outputs = self.model(data)
+                preds = torch.argmax(outputs["edge_type_preds"], dim=1).cpu().numpy()
+                labels = data.edge_types.cpu().numpy()
+                
+                all_preds.extend(preds)
+                all_labels.extend(labels)
+        
+        cm = confusion_matrix(all_labels, all_preds)
+        
+        plt.figure(figsize=(8, 6))
+        sns.heatmap(cm, annot=True, fmt="d", cmap="Blues")
+        plt.xlabel("Predicted Label")
+        plt.ylabel("True Label")
+        plt.title(f"{phase} Confusion Matrix")
+
+        cm_path = f"{self.checkpoint_dir}/{phase}_conf_matrix_epoch.png"
+        plt.savefig(cm_path)
+        plt.close()
+
+        wandb.log({f"{phase}/Confusion_Matrix": wandb.Image(cm_path)})
+
+    def train_step(self, data):
+        raise NotImplementedError("train_step should be implemented in subclasses")
+
+    def validate_step(self, data):
+        raise NotImplementedError("validate_step should be implemented in subclasses")
+
+    def train(self, num_epochs):
+        training_bar = tqdm.tqdm(range(num_epochs))
+        training_bar.set_description("Training")
+
+        for epoch in training_bar:
+            if INTERRUPTED:
+                self.save_checkpoint(epoch, self.best_val_loss, "model_interrupted.pt")
+                training_bar.write("Training interrupted, last model saved.")
+                break
+
+            train_losses = []
+            val_losses = []
+
+            for batch in self.train_loader:
+                train_losses.append(self.train_step(batch))
+
+            for batch in self.val_loader:
+                val_losses.append(self.validate_step(batch))
+
+            avg_train_losses = {k: sum(d[k] for d in train_losses) / len(train_losses) for k in train_losses[0]}
+            avg_val_losses = {k: sum(d[k] for d in val_losses) / len(val_losses) for k in val_losses[0]}
+            avg_val_loss = sum(avg_val_losses.values())
+
+            for k, v in avg_train_losses.items():
+                self.writer.add_scalar(f"Train/{k}", v, epoch)
+                wandb.log({f"Train/{k}": v, "epoch": epoch})
+            for k, v in avg_val_losses.items():
+                self.writer.add_scalar(f"Validation/{k}", v, epoch)
+                wandb.log({f"Validation/{k}": v, "epoch": epoch})
+
+            self.log_histograms(epoch)
+
+            if epoch % 10 == 0:
+                example_data = next(iter(self.train_loader)).to(self.device)
+                self.log_embeddings(epoch, example_data)
+
+            training_bar.write(f"Train Losses: {avg_train_losses}")
+            training_bar.write(f"Val Losses: {avg_val_losses}")
+
+            if avg_val_loss < self.best_val_loss:
+                self.best_val_loss = avg_val_loss
+                self.early_stopping_counter = 0
+                self.save_checkpoint(epoch, avg_val_loss)
+            else:
+                self.early_stopping_counter += 1
+
+            if self.early_stopping_counter >= self.patience:
+                training_bar.write("Early stopping triggered!")
+                break
+        
+        self.writer.close()
+
+    def log_histograms(self, epoch):
+        for name, param in self.model.named_parameters():
+            self.writer.add_histogram(name, param, epoch)
+
+    def log_embeddings(self, epoch, data):
+        self.model.eval()
+        with torch.no_grad():
+            node_embeddings = self.model(data)["node_type_preds"]
+            metadata = [f"Node {i}" for i in range(node_embeddings.size(0))]
+            timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
+            global_step = f"{epoch}_{timestamp}"
+            self.writer.add_embedding(node_embeddings, metadata=metadata, global_step=global_step, tag="embeddings")
+            wandb.log({"embeddings": node_embeddings.cpu().data.numpy(), "epoch": epoch})
+
+
