+++ ApexDAG/nn/training/finetuning_trainer.py
@@ -1,40 +1,14 @@
 import os
-import tqdm
-import wandb
 import torch
-import torch.nn as nn
-import torch.optim as optim
-from datetime import datetime
+
+from ApexDAG.nn.training.base_trainer import BaseTrainer
 from torch_geometric.loader import DataLoader
-from torch.utils.tensorboard import SummaryWriter
 
-# TODO: fix inheriance and introduce a base class for training
-global INTERRUPTED
-INTERRUPTED = False
 
-class FinetuningTrainer:
-    def __init__(self, model, train_dataset, val_dataset, test_dataset, device="cpu", log_dir="runs/", checkpoint_dir="checkpoints/", patience=10):
-        self.model = model.to(device)
-        self.device = device
-        self.train_loader = DataLoader(train_dataset, batch_size=32, shuffle=True)
-        self.val_loader = DataLoader(val_dataset, batch_size=32, shuffle=False)
+class FinetuningTrainer(BaseTrainer):
+    def __init__(self, model, train_dataset, val_dataset, test_dataset, **kwargs):
+        super().__init__(model, train_dataset, val_dataset, **kwargs)
         self.test_loader = DataLoader(test_dataset, batch_size=32, shuffle=False)
-        self.optimizer = optim.Adam(self.model.parameters(), lr=0.001)
-        self.criterion_node = nn.CrossEntropyLoss()
-        self.criterion_edge_type = nn.CrossEntropyLoss()
-        self.criterion_edge_existence = nn.BCELoss()
-        
-        # TensorBoard writer
-        self.writer = SummaryWriter(log_dir=log_dir)
-        
-        # Checkpoint directory
-        self.checkpoint_dir = checkpoint_dir
-        os.makedirs(checkpoint_dir, exist_ok=True)
-        
-        # Early stopping variables
-        self.patience = patience
-        self.best_val_loss = float("inf")
-        self.early_stopping_counter = 0
 
     def save_checkpoint(self, epoch, val_loss, filename=None):
         if filename is None:
@@ -77,69 +51,3 @@ class FinetuningTrainer:
             losses["edge_type_loss"] = self.criterion_edge_type(outputs["edge_type_preds"], data.edge_types)
     
         return {k: v.item() for k, v in losses.items()}
-
-    def train(self, num_epochs):
-        training_bar = tqdm.tqdm(range(num_epochs))
-        training_bar.set_description("Training")
-
-        for epoch in training_bar:
-            is_interrupted = INTERRUPTED
-            if is_interrupted:
-                self.save_checkpoint(epoch, self.best_val_loss, "model_interrupted.pt")
-                training_bar.write("Training interrupted, last model saved.")
-                break
-
-            train_losses = []
-            val_losses = []
-
-            for batch in self.train_loader:
-                train_losses.append(self.train_step(batch))
-
-            for batch in self.val_loader:
-                val_losses.append(self.validate_step(batch))
-
-            avg_train_losses = {k: sum(d[k] for d in train_losses) / len(train_losses) for k in train_losses[0]}
-            avg_val_losses = {k: sum(d[k] for d in val_losses) / len(val_losses) for k in val_losses[0]}
-            avg_val_loss = sum(avg_val_losses.values())
-
-            for k, v in avg_train_losses.items():
-                self.writer.add_scalar(f"Train/{k}", v, epoch)
-                wandb.log({f"Train/{k}": v, "epoch": epoch})
-            for k, v in avg_val_losses.items():
-                self.writer.add_scalar(f"Validation/{k}", v, epoch)
-                wandb.log({f"Validation/{k}": v, "epoch": epoch})
-
-            self.log_histograms(epoch)
-            if epoch % 10 == 0: 
-                example_data = next(iter(self.train_loader)).to(self.device)
-                self.log_embeddings(epoch, example_data)
-
-            training_bar.write(f"Train Losses: {avg_train_losses}")
-            training_bar.write(f"Val Losses: {avg_val_losses}")
-
-            if avg_val_loss < self.best_val_loss:
-                self.best_val_loss = avg_val_loss
-                self.early_stopping_counter = 0
-                self.save_checkpoint(epoch, avg_val_loss)
-            else:
-                self.early_stopping_counter += 1
-
-            if self.early_stopping_counter >= self.patience:
-                training_bar.write("Early stopping triggered!")
-                break
-        
-        self.writer.close()
-
-    def log_histograms(self, epoch):
-        for name, param in self.model.named_parameters():
-            self.writer.add_histogram(name, param, epoch)
-
-    def log_embeddings(self, epoch, data):
-        self.model.eval()
-        with torch.no_grad():
-            node_embeddings = self.model(data)["node_type_preds"]
-            metadata = [f"Node {i}" for i in range(node_embeddings.size(0))]
-            timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-            global_step = f"{epoch}_{timestamp}"
-            self.writer.add_embedding(node_embeddings, metadata=metadata, global_step=global_step, tag="embeddings")
-            wandb.log({"embeddings": node_embeddings.cpu().data.numpy(), "epoch": epoch})
